3.1.7.2 beta

+++ app/src/main/java/net/veldor/flibustaloader/selections/HistoryItem.kt
@@ -2,7 +2,7 @@ package net.veldor.flibustaloader.selections
 
 class HistoryItem(
     val nextPageLink: String?,
-    val clickedItem: Int,
+    val clickedItem: Long,
     val loadedValues: Int,
     val filteredValues: Int,
     val results: ArrayList<FoundedEntity>
